move routing stuff to specific folder
